more unit tests for some utility functions. also moved some functions to utils.js. #5.
